feat(match): add validation PD-2417
