+++ packages/match/controller/src/index.js
@@ -284,7 +284,7 @@ const markupToText = (s) => {
 
 
 export const validate = (model = {}, config = {}) => {
-  const { rows, choiceMode } = model;
+  const { rows, choiceMode, headers } = model;
   const {
     minQuestions,
     maxQuestions,
@@ -294,6 +294,7 @@ export const validate = (model = {}, config = {}) => {
     maxLengthFirstColumnHeading
   } = config;
   const rowsErrors = {};
+  const columnsErrors = {};
   const errors = {};
 
   if (rows.length < minQuestions) {
@@ -302,7 +303,7 @@ export const validate = (model = {}, config = {}) => {
     errors.noOfRowsError = `No more than ${maxQuestions} question rows should be defined.`;
   }
 
-  (rows || []).forEach((row) => {
+  (rows || []).forEach((row, index) => {
     const { id, values = [], title } = row;
     let hasCorrectResponse = false;
 
@@ -312,6 +313,15 @@ export const validate = (model = {}, config = {}) => {
       rowsErrors[id] += `Questions rows content length should not be more than ${maxLengthQuestionsHeading} characters. `;
     } else if (title === '' || title === '<div></div>') {
       rowsErrors[id] += 'Content should not be empty. ';
+    } else {
+      const identicalAnswer = rows.slice(index + 1).some((r) => {
+        console.log('r', r, title);
+        return r.title === title;
+      });
+
+      if (identicalAnswer) {
+        rowsErrors[id] = 'Content should be unique. ';
+      }
     }
 
     values.forEach((value) => {
@@ -325,12 +335,41 @@ export const validate = (model = {}, config = {}) => {
     }
   });
 
+  if (maxAnswers && headers.length - 1 > maxAnswers) {
+    errors.columnsLengthError = `There should be maximum ${maxAnswers} answers.`;
+  }
+
+  if (maxLengthFirstColumnHeading && headers[0].length > maxLengthFirstColumnHeading) {
+    columnsErrors[0] = `The first column heading should have the maximum length of ${maxLengthFirstColumnHeading} characters.`;
+  }
+
+  (headers || []).slice(1).forEach((heading, index) => {
+    if (heading === '' || heading === '<div></div>') {
+      columnsErrors[index + 1] = 'Content should not be empty.';
+    } else if (maxLengthAnswers && heading.length > maxLengthAnswers) {
+      columnsErrors[index + 1] = `Content length should be maximum ${maxLengthAnswers} characters.`;
+    } else {
+      const identicalAnswer = headers.slice(index + 2).some((h) => {
+        return h === heading;
+      });
+
+      if (identicalAnswer) {
+        columnsErrors[index + 2] = 'Content should be unique.';
+      }
+    }
+  });
+
+
   if (!isEmpty(rowsErrors)) {
     errors.rowsErrors = rowsErrors;
     errors.correctResponseError =
       choiceMode === 'radio'
         ? 'There should be a correct response defined for every row.'
         : 'There should be at least one correct response defined for every row.';
+  }
+
+  if (!isEmpty(columnsErrors)) {
+    errors.columnsErrors = columnsErrors;
   }
 
   return errors;
